% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.spacc}
\alias{c.spacc}
\title{Combine spacc Objects}
\usage{
\method{c}{spacc}(...)
}
\arguments{
\item{...}{Named \code{spacc} objects to combine.}
}
\value{
A \code{spacc_multi} object for comparison plotting.
}
\description{
Combine spacc Objects
}
