% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.spacc_beta}
\alias{c.spacc_beta}
\title{Combine spacc_beta Objects}
\usage{
\method{c}{spacc_beta}(...)
}
\arguments{
\item{...}{\code{spacc_beta} objects to combine.}
}
\value{
A combined \code{spacc_beta} object with more seeds.
}
\description{
Combine multiple \code{spacc_beta} objects by stacking their curve matrices.
All objects must have the same number of sites and index.
}
