% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.spacc_coverage}
\alias{c.spacc_coverage}
\title{Combine spacc_coverage Objects}
\usage{
\method{c}{spacc_coverage}(...)
}
\arguments{
\item{...}{\code{spacc_coverage} objects to combine.}
}
\value{
A combined \code{spacc_coverage} object with more seeds.
}
\description{
Combine multiple \code{spacc_coverage} objects by stacking their curve matrices.
All objects must have the same number of sites.
}
