% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.spacc_hill}
\alias{c.spacc_hill}
\title{Combine spacc_hill Objects}
\usage{
\method{c}{spacc_hill}(...)
}
\arguments{
\item{...}{\code{spacc_hill} objects to combine.}
}
\value{
A combined \code{spacc_hill} object with more seeds.
}
\description{
Combine multiple \code{spacc_hill} objects by stacking their curve matrices.
All objects must have the same number of sites.
}
