% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_coverage}
\alias{calc_coverage}
\title{Calculate Good-Turing Coverage Estimate}
\usage{
calc_coverage(abundances)
}
\arguments{
\item{abundances}{Vector of species abundances}
}
\value{
Coverage estimate (0 to 1)
}
\description{
Coverage = proportion of community represented by observed species.
Uses Chao & Jost (2012) estimator.
}
