% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_faith_pd}
\alias{calc_faith_pd}
\title{Calculate Faith's PD from branch lengths}
\usage{
calc_faith_pd(edge, edge_length, n_tips, species_present)
}
\arguments{
\item{edge}{Integer matrix (n_edges x 2) of tree edges}

\item{edge_length}{Numeric vector of edge lengths}

\item{n_tips}{Number of tips in tree}

\item{species_present}{Logical vector indicating which tips are present}
}
\value{
Faith's PD
}
\description{
Sum of branch lengths connecting species in sample to root.
Requires edge matrix and edge lengths from ape::phylo object.
}
