% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_fdis}
\alias{calc_fdis}
\title{Calculate Functional Dispersion (FDis)}
\usage{
calc_fdis(traits, species_present, abundances)
}
\arguments{
\item{traits}{Numeric matrix (species x traits)}

\item{species_present}{Logical vector of present species}

\item{abundances}{Abundance vector}
}
\value{
FDis value
}
\description{
Mean distance to abundance-weighted centroid in trait space.
}
