% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_hill_number}
\alias{calc_hill_number}
\title{Calculate Hill Number for a vector of abundances}
\usage{
calc_hill_number(abundances, q)
}
\arguments{
\item{abundances}{Vector of species abundances}

\item{q}{Order of diversity (0 = richness, 1 = Shannon, 2 = Simpson)}
}
\value{
Hill number (effective number of species)
}
\description{
Calculate Hill Number for a vector of abundances
}
