% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_mntd}
\alias{calc_mntd}
\title{Calculate Mean Nearest Taxon Distance (MNTD)}
\usage{
calc_mntd(
  dist_mat,
  species_present,
  abundance_weighted = FALSE,
  abundances = numeric()
)
}
\arguments{
\item{dist_mat}{Pairwise distance matrix (species x species)}

\item{species_present}{Logical vector indicating which species are present}

\item{abundance_weighted}{If true, weight by abundances}

\item{abundances}{Vector of abundances}
}
\value{
MNTD value
}
\description{
Mean distance to nearest relative for each species in sample.
}
