% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_mpd}
\alias{calc_mpd}
\title{Calculate Mean Pairwise Distance (MPD)}
\usage{
calc_mpd(
  dist_mat,
  species_present,
  abundance_weighted = FALSE,
  abundances = numeric()
)
}
\arguments{
\item{dist_mat}{Pairwise distance matrix (species x species)}

\item{species_present}{Logical vector indicating which species are present}

\item{abundance_weighted}{If true, weight by abundances}

\item{abundances}{Vector of abundances (used if abundance_weighted = true)}
}
\value{
MPD value
}
\description{
Mean phylogenetic or functional distance between species in a sample.
}
