% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{chao1}
\alias{chao1}
\title{Chao1 Richness Estimator}
\usage{
chao1(x)
}
\arguments{
\item{x}{A site-by-species matrix (abundance data). Columns are pooled
across sites.}
}
\value{
An object of class \code{spacc_estimate} with components:
\item{estimator}{Name of the estimator (\code{"chao1"})}
\item{estimate}{Estimated total richness}
\item{se}{Standard error of the estimate}
\item{lower}{Lower 95\\% confidence bound}
\item{upper}{Upper 95\\% confidence bound}
\item{S_obs}{Observed species richness}
\item{details}{List with \code{f1} (singletons) and \code{f2} (doubletons)}
}
\description{
Estimate total species richness from abundance data using the Chao1
estimator (Chao 1984). Uses the number of singletons and doubletons
to estimate undetected species.
}
\details{
The Chao1 estimator is:
\deqn{S_{Chao1} = S_{obs} + \frac{f_1^2}{2 f_2}}
where \eqn{f_1} is the number of singletons (species with total abundance 1)
and \eqn{f_2} is the number of doubletons (abundance 2).

When \eqn{f_2 = 0}, the bias-corrected form is used:
\deqn{S_{Chao1} = S_{obs} + \frac{f_1 (f_1 - 1)}{2}}
}
\examples{
species <- matrix(rpois(50 * 30, 2), nrow = 50)
chao1(species)

}
\references{
Chao, A. (1984). Nonparametric estimation of the number of classes in a
population. Scandinavian Journal of Statistics, 11, 265-270.

Chao, A. (1987). Estimating the population size for capture-recapture data
with unequal catchability. Biometrics, 43, 783-791.
}
\seealso{
\code{\link[=chao2]{chao2()}} for incidence-based estimation, \code{\link[=ace]{ace()}} for
abundance-based coverage estimator
}
