% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{chao2}
\alias{chao2}
\title{Chao2 Richness Estimator}
\usage{
chao2(x)
}
\arguments{
\item{x}{A site-by-species matrix (presence/absence or abundance).
Automatically binarized.}
}
\value{
An object of class \code{spacc_estimate}.
}
\description{
Estimate total species richness from incidence (presence/absence) data
using the Chao2 estimator (Chao 1987).
}
\details{
The Chao2 estimator is the incidence-based analogue of Chao1:
\deqn{S_{Chao2} = S_{obs} + \frac{Q_1^2}{2 Q_2}}
where \eqn{Q_1} is the number of uniques (species found at exactly 1 site)
and \eqn{Q_2} is the number of duplicates (species found at exactly 2 sites).
}
\examples{
species <- matrix(rbinom(50 * 30, 1, 0.3), nrow = 50)
chao2(species)

}
\references{
Chao, A. (1987). Estimating the population size for capture-recapture data
with unequal catchability. Biometrics, 43, 783-791.
}
\seealso{
\code{\link[=chao1]{chao1()}} for abundance-based estimation
}
