% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical.R
\name{collector}
\alias{collector}
\title{Collector's Curve}
\usage{
collector(x)
}
\arguments{
\item{x}{A site-by-species matrix.}
}
\value{
A data.frame with columns: sites, species
}
\description{
Compute the species accumulation curve in the order sites appear
in the data (no randomization). Useful for understanding how data
was collected.
}
