% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_beta_knn_parallel}
\alias{cpp_beta_knn_parallel}
\title{Parallel kNN Beta Diversity Accumulation}
\usage{
cpp_beta_knn_parallel(
  species_pa,
  dist_mat,
  n_seeds,
  use_jaccard = FALSE,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{dist_mat}{Distance matrix}

\item{n_seeds}{Number of starting points}

\item{use_jaccard}{Use Jaccard instead of Sorensen}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
List with matrices for each beta component
}
\description{
Parallel kNN Beta Diversity Accumulation
}
