% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_collector_single}
\alias{cpp_collector_single}
\title{Collector Accumulation (data order)}
\usage{
cpp_collector_single(species_pa)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}
}
\value{
Integer vector of cumulative species counts
}
\description{
Collector Accumulation (data order)
}
