% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cone_parallel}
\alias{cpp_cone_parallel}
\title{Parallel Directional Cone Accumulation}
\usage{
cpp_cone_parallel(
  species_pa,
  x,
  y,
  n_seeds,
  width = 0.785398,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}

\item{n_seeds}{Number of random starting points}

\item{width}{Cone half-width in radians (default pi/4)}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
Integer matrix of accumulation curves
}
\description{
Parallel Directional Cone Accumulation
}
