% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_distance_decay_parallel}
\alias{cpp_distance_decay_parallel}
\title{Parallel Distance-Decay Accumulation}
\usage{
cpp_distance_decay_parallel(
  species_pa,
  dist_mat,
  n_seeds,
  breaks,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{dist_mat}{Numeric matrix of pairwise distances}

\item{n_seeds}{Number of random starting points}

\item{breaks}{Distance thresholds}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
Integer matrix of species counts
}
\description{
Parallel Distance-Decay Accumulation
}
