% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_func_knn_parallel}
\alias{cpp_func_knn_parallel}
\title{Parallel Functional Diversity Accumulation}
\usage{
cpp_func_knn_parallel(
  species_mat,
  site_dist_mat,
  traits,
  n_seeds,
  metrics,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_mat}{Integer matrix (sites x species) abundances}

\item{site_dist_mat}{Site distance matrix}

\item{traits}{Trait matrix (species x traits)}

\item{n_seeds}{Number of random starting points}

\item{metrics}{Character vector of metrics to compute}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
List with matrices for each metric
}
\description{
Parallel Functional Diversity Accumulation
}
