% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_func_knn_single}
\alias{cpp_func_knn_single}
\title{Single kNN Functional Diversity Accumulation}
\usage{
cpp_func_knn_single(species_mat, site_dist_mat, traits, seed, metrics)
}
\arguments{
\item{species_mat}{Integer matrix (sites x species) abundances}

\item{site_dist_mat}{Site distance matrix}

\item{traits}{Trait matrix (species x traits)}

\item{seed}{Starting site}

\item{metrics}{Metrics: "fdis", "fric"}
}
\value{
List with curves for each metric
}
\description{
Single kNN Functional Diversity Accumulation
}
