% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_gaussian_parallel}
\alias{cpp_gaussian_parallel}
\title{Parallel Gaussian-weighted Accumulation}
\usage{
cpp_gaussian_parallel(
  species_pa,
  dist_mat,
  n_seeds,
  sigma,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{dist_mat}{Numeric matrix of pairwise distances}

\item{n_seeds}{Number of random starting points}

\item{sigma}{Gaussian bandwidth parameter}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
Integer matrix of accumulation curves
}
\description{
Parallel Gaussian-weighted Accumulation
}
