% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_kncn_parallel}
\alias{cpp_kncn_parallel}
\title{Parallel kNCN Accumulation}
\usage{
cpp_kncn_parallel(species_pa, x, y, n_seeds, n_cores = 1L, progress = FALSE)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}

\item{n_seeds}{Number of random starting points}

\item{n_cores}{Number of cores to use}

\item{progress}{Show progress (currently ignored)}
}
\value{
Integer matrix (n_seeds x n_sites) of accumulation curves
}
\description{
Run kNCN accumulation from multiple random starting points in parallel.
}
