% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_knn_coverage_parallel}
\alias{cpp_knn_coverage_parallel}
\title{Parallel kNN Accumulation with Coverage}
\usage{
cpp_knn_coverage_parallel(
  species_mat,
  dist_mat,
  n_seeds,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_mat}{Integer matrix (sites x species)}

\item{dist_mat}{Distance matrix}

\item{n_seeds}{Number of starting points}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
List with richness, individuals, coverage matrices
}
\description{
Parallel kNN Accumulation with Coverage
}
