% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_knn_parallel}
\alias{cpp_knn_parallel}
\title{Parallel kNN Accumulation}
\usage{
cpp_knn_parallel(species_pa, dist_mat, n_seeds, n_cores = 1L, progress = FALSE)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{dist_mat}{Numeric matrix of pairwise distances}

\item{n_seeds}{Number of random starting points}

\item{n_cores}{Number of cores to use}

\item{progress}{Show progress (currently ignored in C++)}
}
\value{
Integer matrix (n_seeds x n_sites) of accumulation curves
}
\description{
Run kNN accumulation from multiple random starting points in parallel.
}
