% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_phylo_knn_parallel}
\alias{cpp_phylo_knn_parallel}
\title{Parallel Phylogenetic Diversity Accumulation}
\usage{
cpp_phylo_knn_parallel(
  species_pa,
  site_dist_mat,
  phylo_dist_mat,
  n_seeds,
  metrics,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{site_dist_mat}{Site distance matrix}

\item{phylo_dist_mat}{Phylogenetic distance matrix}

\item{n_seeds}{Number of starting points}

\item{metrics}{Metrics to compute}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
List with matrices for each metric
}
\description{
Parallel Phylogenetic Diversity Accumulation
}
