% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_phylo_knn_single}
\alias{cpp_phylo_knn_single}
\title{Single kNN Phylogenetic Diversity Accumulation}
\usage{
cpp_phylo_knn_single(species_pa, site_dist_mat, phylo_dist_mat, seed, metrics)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{site_dist_mat}{Site distance matrix}

\item{phylo_dist_mat}{Phylogenetic/functional distance matrix (species x species)}

\item{seed}{Starting site}

\item{metrics}{Character vector of metrics: "mpd", "mntd"}
}
\value{
List with accumulation curves for each metric
}
\description{
Single kNN Phylogenetic Diversity Accumulation
}
