% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_radius_parallel}
\alias{cpp_radius_parallel}
\title{Parallel Radius-Order Accumulation}
\usage{
cpp_radius_parallel(
  species_pa,
  dist_mat,
  n_seeds,
  n_cores = 1L,
  progress = FALSE
)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{dist_mat}{Numeric matrix of pairwise distances}

\item{n_seeds}{Number of random starting points}

\item{n_cores}{Number of cores}

\item{progress}{Show progress}
}
\value{
Integer matrix of accumulation curves
}
\description{
Parallel Radius-Order Accumulation
}
