% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_random_parallel}
\alias{cpp_random_parallel}
\title{Parallel Random Accumulation}
\usage{
cpp_random_parallel(species_pa, n_seeds, n_cores = 1L, progress = FALSE)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species)}

\item{n_seeds}{Number of random permutations}

\item{n_cores}{Number of cores to use}

\item{progress}{Show progress (currently ignored)}
}
\value{
Integer matrix (n_seeds x n_sites) of accumulation curves
}
\description{
Run random-order accumulation multiple times (null model).
}
