% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_random_single}
\alias{cpp_random_single}
\title{Single Random Accumulation Curve}
\usage{
cpp_random_single(species_pa, order)
}
\arguments{
\item{species_pa}{Integer matrix (sites x species), presence/absence}

\item{order}{Integer vector giving the order to visit sites (0-based)}
}
\value{
Integer vector of cumulative species counts
}
\description{
Visit sites in random order (null model for comparison).
}
