% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate_at_coverage}
\alias{interpolate_at_coverage}
\title{Interpolate Richness at Target Coverage Levels}
\usage{
interpolate_at_coverage(richness, coverage, targets)
}
\arguments{
\item{richness}{Vector of cumulative richness}

\item{coverage}{Vector of coverage values}

\item{targets}{Target coverage levels to interpolate}
}
\value{
Vector of interpolated richness values
}
\description{
Interpolate Richness at Target Coverage Levels
}
