% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{c.spacc}
\alias{c.spacc}
\title{Combine spacc Objects}
\usage{
\method{c}{spacc}(...)
}
\arguments{
\item{...}{Named \code{spacc} objects to combine into a grouped \code{spacc}.}
}
\value{
A grouped \code{spacc} object with per-group curves.
}
\description{
Combine spacc Objects
}
