% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spacc-package}
\alias{spacc-package}
\title{spacc: Fast Spatial Species Accumulation Curves}
\description{
High-performance spatial species accumulation curves using nearest-neighbor algorithms. Implements 'kNN' and 'kNCN' sampling methods with a 'C++' backend for speed. Supports Hill numbers (q=0,1,2), beta diversity partitioning (turnover/nestedness), coverage-based rarefaction and extrapolation, phylogenetic diversity (Faith's PD, mean pairwise distance, mean nearest taxon distance), functional diversity accumulation, diversity-area relationships (DAR), endemism-area curves, sampling-effort correction and fragmentation analysis, species-area relationship (SAR) models based on extreme value theory (EVT), and area-based richness extrapolation via the total-species curve. Multiple starting points (seeds) provide uncertainty quantification. Methods are described in 'Chao' et al. (2014) \doi{10.1890/13-0133.1}, 'Baselga' (2010) \doi{10.1111/j.1466-8238.2009.00490.x}, 'Chao' and 'Jost' (2012) \doi{10.1890/11-1952.1}, 'Faith' (1992) \doi{10.1016/0006-3207(92)91201-3}, 'Ma' (2018) \doi{10.1002/ece3.4526}, 'Borda-de-Agua' et al. (2025) \doi{10.1038/s41467-025-59239-7}, 'Hanski' et al. (2013) \doi{10.1073/pnas.1311190110}, 'Jost' (2007) \doi{10.1890/06-1736.1}, and 'Ugland' et al. (2003) \doi{10.1046/j.1365-2656.2003.00748.x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gillescolling.com/spacc/}
  \item \url{https://github.com/gcol33/spacc}
  \item Report bugs at \url{https://github.com/gcol33/spacc/issues}
}

}
\author{
\strong{Maintainer}: Gilles Colling \email{gilles.colling051@gmail.com} (\href{https://orcid.org/0000-0003-3070-6066}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Gilles Colling \email{gilles.colling051@gmail.com} (\href{https://orcid.org/0000-0003-3070-6066}{ORCID}) [copyright holder]
}

}
\keyword{internal}
