% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spacc-package}
\alias{spacc-package}
\title{spacc: Fast Spatial Species Accumulation Curves}
\description{
High-performance spatial species accumulation curves using nearest-neighbor algorithms. Implements kNN and kNCN sampling methods with a C++ backend for speed. Supports Hill numbers (q=0,1,2), beta diversity partitioning (turnover/nestedness), coverage-based rarefaction, and phylogenetic/functional diversity accumulation. Multiple starting points (seeds) provide uncertainty quantification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gillescolling.com/spacc/}
  \item \url{https://github.com/gcol33/spacc}
  \item Report bugs at \url{https://github.com/gcol33/spacc/issues}
}

}
\author{
\strong{Maintainer}: Gilles Colling \email{gilles.colling051@gmail.com} (\href{https://orcid.org/0000-0003-3070-6066}{ORCID}) [copyright holder]

}
\keyword{internal}
